Fixed binaries in pypi distribution (#48)
Added directory pysmt/cmd for commands/binaries to distribute. This
makes the importing of these "utilities" cleaner, since now they are
within the pysmt module. I.e., pysmt.cmd.install .
